feat: Possibilty to remove devFormat and hence node specific libraries
It is not possible to alias dev-format to prod-format in webpack to
remove node specific code
